Refactor useSquareLayoutItems to use ResizeObserver API, add layout styling to Main view
