+++ src/assets/styles/styles.scss
@@ -9,6 +9,8 @@
 }
 
 body {
+	margin: 0;
+	padding: 0;
 	font-family: OpenSans, Tahoma, Geneva, Verdana, sans-serif;
 }
