+++ src/assets/styles/variables.scss
@@ -1,3 +1,6 @@
+$page-content-vertical-padding: 10px;
+$page-content-horizontal-padding: 15px;
+
 $input-background-color: #fff;
 $input-border-color: #999;
 $input-border-radius: 6px;
