+++ src/views/Main/main.module.scss
@@ -1,0 +1,22 @@
+@import '../../assets/styles/variables';
+
+.page-content {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+
+	padding: $page-content-vertical-padding $page-content-horizontal-padding;
+
+	@include desktop-styling {
+		padding-left: $sidebar-covered-width + $page-content-horizontal-padding;
+		padding-right: $page-content-horizontal-padding;
+	}
+}
+
+.bookmarks-layout {
+	width: 100%;
+
+	@include desktop-styling {
+		width: Min(90%, 700px) !important;
+	}
+}
